More refactoring and work on tree boss
Changed the code up a bit more becuase of some bugs in the tutorial. Blank boss made to avoid such bugs. Started working on the walk animation for the tree but its saying "can't read index of undefined", even though the boss isn't undefined. Rather it's been defined thrice by the point you reach level 2. Plan to work on it