Fixing boss
Flipped the evil heart, it was in the wrong orientation. Also sort of fixed the giant, made him walk left when knight overlaps with him. This messed up the tree though. Lastly, tree projectile now disappear when colliding with platforms. not sure if we want to keep this, as it'll make the game easier than it already is