Tutorial jump and attack fixes
Can jump in tutorial again. hitPlatform wasn't assigned anything in tutorial nor was created in knight.js, so did that and it fixed the problem

Got older code from a past sprint to disallow spamming of attack button by having a timer between each attack. Also knightBox is on time now with the attack animation.